+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/storage/query/ISpanAttachedEventQueryDAO.java
@@ -1,0 +1,11 @@
+package org.apache.skywalking.oap.server.core.storage.query;
+
+import org.apache.skywalking.oap.server.core.analysis.manual.spanattach.SpanAttachedEventRecord;
+import org.apache.skywalking.oap.server.library.module.Service;
+
+import java.io.IOException;
+import java.util.List;
+
+public interface ISpanAttachedEventQueryDAO extends Service {
+    List<SpanAttachedEventRecord> querySpanAttachedEvents(String traceId) throws IOException;
+}